+++ lib/EpiNetwork/test/runtests.jl
@@ -10,3 +10,5 @@ using Graphs
     include("test_graphs_ext.jl")
     include("test_network_likelihood.jl")
 end
+
+include("test_actions.jl")